requirements.txt
aiogram>=3.3.0
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
google-cloud-firestore==2.14.0
aiohttp==3.9.3
pydantic>=2.4.1,<2.6.0
pydantic-settings>=2.1.0,<3.0.0
python-json-logger==2.0.7
APScheduler==3.10.4 